requirements.txt
flask
pandas
numpy
sklearn
# -e . #this commands execute the housing package
ipykernel
six
PyYaml